Move TestingKafka to presto-testing-kafka module
TestingKafka can be used by multiple modules.
