+++ presto-kafka/src/test/java/io/prestosql/plugin/kafka/TestKafkaDistributed.java
@@ -13,9 +13,9 @@
  */
 package io.prestosql.plugin.kafka;
 
-import io.prestosql.plugin.kafka.util.TestingKafka;
 import io.prestosql.testing.AbstractTestQueries;
 import io.prestosql.testing.QueryRunner;
+import io.prestosql.testing.kafka.TestingKafka;
 import io.prestosql.tpch.TpchTable;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.Test;
