+++ presto-testing-kafka/src/main/java/io/prestosql/testing/kafka/TestingKafka.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.plugin.kafka.util;
+package io.prestosql.testing.kafka;
 
 import com.google.common.collect.ImmutableMap;
 import org.apache.kafka.clients.producer.KafkaProducer;
